+++ Sources/PulseUI/Features/Inspector/NetworkInspectorView-watchos.swift
@@ -12,7 +12,7 @@ import Combine
 struct NetworkInspectorView: View {
     @ObservedObject var task: NetworkTaskEntity
 
-    @State private var isCurrentRequest = false
+    @ObservedObject private var settings: UserSettings = .shared
 
     var body: some View {
         contents
@@ -31,8 +31,8 @@ struct NetworkInspectorView: View {
             }
             Section {
                 makeTransferInfo(isReceivedHidden: true)
-                NetworkInspectorRequestTypePicker(isCurrentRequest: $isCurrentRequest)
-                NetworkInspectorView.makeRequestSection(task: task, isCurrentRequest: isCurrentRequest)
+                NetworkInspectorRequestTypePicker(isCurrentRequest: $settings.isShowingCurrentRequest)
+                NetworkInspectorView.makeRequestSection(task: task, isCurrentRequest: settings.isShowingCurrentRequest)
             }
             if task.state != .pending {
                 Section {